chore: package frontend properly with Nix
This fixes what previous commits broke; we can build the frontend and
backend independently, as well as (ideally) use the NixOS modlue. (I
haven't tested this.)

Note that we hardcode the randomly selected backend port: I don't know
when it is that vite.config.ts 'runs,' so I'm not sure if we can have it
load the port from an environment variable.